+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### NEXT
 
+- Add `jitter` in `Consumer` 'outbound-rtp' stats ([PR #1654](https://github.com/versatica/mediasoup/pull/1654)).
+
 ### 3.19.9
 
 - Fix RTCP packets lost in stats ([PR #1651](https://github.com/versatica/mediasoup/pull/1651)).