+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ### NEXT
 
-Fix RTX packets containing non yet seen RTP packets being discarded ([PR #1653](https://github.com/versatica/mediasoup/pull/1653), credits to @penguinol and @mengbieting).
+- Fix RTX packets containing non yet seen RTP packets being discarded ([PR #1653](https://github.com/versatica/mediasoup/pull/1653), credits to @penguinol and @mengbieting).
+
+### 3.19.10
+
+- Add `jitter` in `Consumer` 'outbound-rtp' stats ([PR #1654](https://github.com/versatica/mediasoup/pull/1654)).
 
 ### 3.19.9